factored out WS class
