+++ src/transports/courier.ts
@@ -1,4 +1,6 @@
+import { WS } from "../ws";
 import { Transport } from "./base";
+import { Intercept } from "./types";
 const LAMBDA_WS_URL = "wss://zj8xquqj55.execute-api.us-east-1.amazonaws.com/dev";
 interface ITransportOptions {
   clientKey: string;
@@ -10,6 +12,7 @@ export class CourierTransport extends Transport {
   protected ws: WS;
   protected clientKey;
   protected secretKey;
+  protected interceptor;
   constructor(options: ITransportOptions) {
     super();
 
@@ -18,7 +21,7 @@ export class CourierTransport extends Transport {
     }
     this.clientKey = options.clientKey;
     this.secretKey = options.secretKey;
-    this.ws = new WS();
+    this.ws = new WS({ url: LAMBDA_WS_URL });
     this.authenticate();
     this.ws.connect(options.clientKey);
   }
@@ -52,77 +55,8 @@ export class CourierTransport extends Transport {
   unsubscribe(channel: string, event: string){
     this.ws.unsubscribe(channel, event, this.clientKey);
   }
-}
 
-export class WS {
-  connection: WebSocket;
-  protected connected;
-  protected messageCallback;
-  protected clientKey;
-  constructor() {
-    this.messageCallback = null;
-    this.connection = null;
-    this.connected = false;
-  }
-  connect(clientKey){
-    this.clientKey = clientKey;
-    const url = `${LAMBDA_WS_URL}/?clientKey=${clientKey}`;
-    this.connection = new WebSocket(url);
-    this.initiateListener();
-  }
-  onMessage({ data }){
-    try {
-      data = JSON.parse(data);
-    } catch {
-      //
-    }
-
-    if (data && this.messageCallback) {
-      this.messageCallback({ data });
-    }
-  }
-  onConnectionOpen(){
-    this.connected = true;
-  }
-  waitForOpen(): Promise<any> {
-    return new Promise(resolve => {
-      if (this.connected) {
-        resolve(true);
-      } else {
-        this.connection.addEventListener("open", resolve);
-      }
-    });
-  }
-  async subscribe(channel, event, clientKey, callback){
-    await this.waitForOpen();
-    this.send({
-      action: "subscribe",
-      data: {
-        channel,
-        event,
-        clientKey,
-      },
-    });
-    this.messageCallback = callback;
-  }
-  send(message){
-    this.connection.send(JSON.stringify(message));
-  }
-  unsubscribe(channel, event, clientKey){
-    this.send({
-      action: "unsubscribe",
-      data: {
-        channel,
-        event,
-        clientKey,
-      },
-    });
-  }
-  close(){
-    this.connection.close();
-  }
-  initiateListener(){
-    this.connection.onopen = this.onConnectionOpen.bind(this);
-    this.connection.onmessage = this.onMessage.bind(this);
+  intercept(cb: Intercept){
+    this.interceptor = cb;
   }
 }
