+++ src/transports/types.ts
@@ -9,3 +9,6 @@ export interface ICourierEvent {
   type: "message",
   data: IMessage,
 }
+
+// eslint-disable-next-line no-unused-vars
+export type Intercept = (message: IMessage) => IMessage | undefined;
